core: Fix typecast of config value VdsRefreshRate
Previous patch[1] changed types of multiple config values
from Integer to Long. This patch fixes one type cast that
was missed.

[1] - https://gerrit.ovirt.org/#/c/81718/

Bug-Url: https://bugzilla.redhat.com/1566112

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/RunVmCommandBase.java
@@ -364,9 +364,9 @@ public abstract class RunVmCommandBase<T extends VmOperationParameterBase> exten
                 .mapToLong(vdsId -> resourceManager.getVdsManager(vdsId).getLastUpdateElapsed())
                 .max().getAsLong();
 
-          long maxWaitTime = Math.min(
+        long maxWaitTime = Math.min(
                 Math.max(maxUpdateElapsed,
-                        TimeUnit.SECONDS.toMillis(Config.<Integer> getValue(VdsRefreshRate))),
+                        TimeUnit.SECONDS.toMillis(Config.<Long> getValue(VdsRefreshRate))),
                 Config.<Integer> getValue(ConfigValues.ThrottlerMaxWaitForVdsUpdateInMillis));
 
         long endTime = System.currentTimeMillis() + maxWaitTime;
